+++ abe/auth/__init__.py
@@ -1,6 +1,6 @@
 """This module defines a decorator to control access to routing methods."""
 # flake8: noqa: F401 F403
 
-from .auth_tokens import *
+from .access_tokens import *
 from .auth_requests import *
 from .auth_decorators import *
